completed tests for entity relations

+++ lib/tekeya/entity.rb
@@ -12,17 +12,13 @@ module Tekeya
       has_many :activities, as: :entity, class_name: ::Tekeya::Activity, dependent: :destroy
     end
 
-    module ClassMethods
-      # A method to identify the entity's heritage
-      def is_entity?
-        return true
-      end
-    end
-
     # Tracks the given entity and copies it's recent feed to the tracker feed
     #
     # @param [Entity] entity the entity to track
     def track(entity)
+      check_if_entity(entity)
+      raise ::Tekeya::Errors::TekeyaRelationAlreadyExists.new("Already tracking #{entity}") if self.tracks?(entity)
+
       add_relation(self, entity, :tracks)
       ::Resque.enqueue(::Tekeya::Feed::Resque::FeedCopy, entity.profile_feed_key, self.feed_key)
     end
@@ -32,7 +28,7 @@ module Tekeya
     # @param  [String, nil] type used to return a certain type of entities being tracked
     # @return [Array] the entities tracked by this entity
     def tracking(type = nil)
-      relations_of(type, :tracks, self)
+      relations_of(self, :tracks, type)
     end
 
     # Returns a list of entities tracking this entity
@@ -40,7 +36,7 @@ module Tekeya
     # @param  [String, nil] type used to return a certain type of entities being tracked
     # @return [Array] the entities tracking this entity
     def trackers(type = nil)
-      relations_of(self, :tracks, type)
+      relations_of(self, :tracks, type, true)
     end
 
     # Checks if this entity is tracking the given entity
@@ -48,6 +44,7 @@ module Tekeya
     # @param  [Entity] entity the entity to check
     # @return [Boolean] true if this entity is tracking the given entity, false otherwise
     def tracks?(entity)
+      check_if_entity(entity)
       relation_exists?(self, entity, :tracks)
     end
 
@@ -55,11 +52,22 @@ module Tekeya
     #
     # @param [Entity] entity the entity to untrack
     def untrack(entity)
+      check_if_entity(entity)
+      raise ::Tekeya::Errors::TekeyaRelationNonExistent.new("Can't untrack an untracked entity") unless self.tracks?(entity)
+
       delete_relation(self, entity, :tracks)
       ::Resque.enqueue(::Tekeya::Feed::Resque::DeleteFeed, entity.profile_feed_key, self.feed_key)
     end
 
     def block(entity)
+      check_if_entity(entity)
+      raise ::Tekeya::Errors::TekeyaRelationAlreadyExists.new("Already blocking #{entity}") if self.blocks?(entity)
+
+      unless entity.is_tekeya_group?
+        self.untrack(entity) if self.tracks?(entity)
+        entity.untrack(self) if entity.tracks?(self)
+      end
+
       add_relation(self, entity, :blocks)
     end
 
@@ -68,15 +76,22 @@ module Tekeya
     end
 
     def blocks?(entity)
+      check_if_entity(entity)
       relation_exists?(self, entity, :blocks)
     end
 
     def unblock(entity)
+      check_if_entity(entity)
+      raise ::Tekeya::Errors::TekeyaRelationNonExistent.new("Can't unblock an unblocked entity") unless self.blocks?(entity)
+
       delete_relation(self, entity, :blocks)
     end
 
     def join(group)
-      add_relation(self, entity, :joins)
+      check_if_group(group)
+      raise ::Tekeya::Errors::TekeyaRelationAlreadyExists.new("Already a member of #{group}") if self.member_of?(group)
+
+      add_relation(self, group, :joins)
     end
 
     def groups(type = nil)
@@ -84,11 +99,15 @@ module Tekeya
     end
 
     def member_of?(group)
-      relation_exists?(self, entity, :joins)
+      check_if_group(group)
+      relation_exists?(self, group, :joins)
     end
 
     def leave(group)
-      delete_relation(self, entity, :joins)
+      check_if_group(group)
+      raise ::Tekeya::Errors::TekeyaRelationNonExistent.new("Can't leave an unjoined group") unless self.member_of?(group)
+
+      delete_relation(self, group, :joins)
     end
 
     def profile_feed
@@ -105,26 +124,50 @@ module Tekeya
       "#{self.class.name}:#{self.send(self.entity_primary_key)}:feed"
     end
 
+    # A method to identify the entity's heritage
+    def is_tekeya_entity?
+      return true
+    end
+
+    def is_tekeya_group?
+      return false
+    end
+
     private
 
+    def check_if_entity(entity)
+      raise ::Tekeya::Errors::TekeyaNonEntity.new("Supplied argument is not a Tekeya::Entity") unless entity.present? && entity.is_tekeya_entity?
+    end
+
+    def check_if_group(group)
+      raise ::Tekeya::Errors::TekeyaNonGroup.new("Supplied argument is not a Tekeya::Entity::Group") unless group.present? && group.is_tekeya_group?
+    end
+
     def add_relation(from, to, type)
-      ::Tekeya.relations.add(from.send(from.class.entity_primary_key), from.class.name, to.send(to.class.entity_primary_key), to.class.name, type)
+      ::Tekeya.relations.add(from.send(from.class.entity_primary_key), from.class.name, to.send(to.class.entity_primary_key), to.class.name, 0, type)
     end
 
     def delete_relation(from, to, type)
-      ::Tekeya.relations.delete(from.send(from.class.entity_primary_key), from.class.name, to.send(to.class.entity_primary_key), to.class.name, :tracks)
+      ::Tekeya.relations.delete(from.send(from.class.entity_primary_key), from.class.name, to.send(to.class.entity_primary_key), to.class.name, type)
     end
 
-    def relations_of(from, relation_type, entity_type)
+    def relations_of(from, relation_type, entity_type, reverse = false)
       result_entity_class = entity_type.constantize if entity_type
-      ::Tekeya.relations.where(from.send(from.class.entity_primary_key), from.class.name, nil, entity_type, relation_type).entries.map do |entry|
-        result_entity_class ||= entry.toEntityType.constantize
-        result_entity_class.where(result_entity_class.entity_primary_key.to_sym => entry.toEntityId).first
+      unless reverse
+        ::Tekeya.relations.where(from.send(from.class.entity_primary_key), from.class.name, nil, entity_type, relation_type).entries.map do |entry|
+          result_entity_class ||= entry.toEntityType.constantize
+          result_entity_class.where(result_entity_class.entity_primary_key.to_sym => entry.toEntityId).first
+        end
+      else
+        ::Tekeya.relations.where(nil, entity_type, from.send(from.class.entity_primary_key), from.class.name, relation_type).entries.map do |entry|
+          result_entity_class ||= entry.fromEntityType.constantize
+          result_entity_class.where(result_entity_class.entity_primary_key.to_sym => entry.fromEntityId).first
+        end
       end
     end
 
     def relation_exists?(from, to, type)
-      !::Tekeya.relations.where(from.send(from.class.entity_primary_key), from.class.name, to.send(to.class.entity_primary_key), entity.class.name, :tracks).entries.empty?
+      !::Tekeya.relations.where(from.send(from.class.entity_primary_key), from.class.name, to.send(to.class.entity_primary_key), to.class.name, type).entries.empty?
     end
   end
 end
